export generic selector into its own file
